requirements.txt
optuna
pandas==2.0.3
torch_geometric==2.6.0
tqdm
moz-sql-parser==4.40.21126
psycopg2
scikit-learn==1.3.2
xgboost==2.1.1
dgl-cu110==0.6.1
mysql-connector